apps/backend/pyproject.toml
[project]
name = "family-tree-backend"
version = "0.1.0"
description = "Backend for the Family Tree project"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "aiosmtplib==1.1.7",
    "alembic==1.11.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cryptography==46.0.1",
    "dnspython==2.8.0",
    "email-validator==2.0.0.post2",
    "fastapi==0.118.0",
    "frozenlist==1.7.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "mailtrap==2.2.0",
    "Mako==1.3.10",
    "MarkupSafe==3.0.3",
    "multidict==6.6.4",
    "propcache==0.3.2",
    "psycopg==3.2.10",
    "psycopg-binary==3.2.10",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic_core==2.33.2",
    "PyJWT==2.10.1",
    "python-dotenv==1.0.0",
    "PyYAML==6.0.3",
    "requests==2.32.5",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.19",
    "starlette==0.48.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.15.0",
    "urllib3==2.5.0",
    "uvicorn==0.23.2",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "Werkzeug==3.1.3",
    "yarl==1.20.1",
]

[tool.setuptools]
packages = ["api", "models", "schemas", "services", "utils", "migrations"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
